Moved exception interfaces to dedicated namespace
Also fixed some docs and uses. Tests also fixed to match.
